+++ desktop/packages/mullvad-vpn/src/shared/routes.ts
@@ -34,4 +34,5 @@ export enum RoutePath {
   manageDevices = '/settings/manage-devices',
   antiCensorship = '/settings/advanced/anti-censorship',
   wireguardPort = '/settings/advanced/wireguard-port',
+  lwo = '/settings/advanced/lwo',
 }